added views for roster and player
